+++ src/main/groovy/life/qbic/datamodel/dtos/business/ProductId.groovy
@@ -1,6 +1,8 @@
 package life.qbic.datamodel.dtos.business
 
 import groovy.transform.EqualsAndHashCode
+import life.qbic.datamodel.dtos.business.services.ProductType
+import life.qbic.datamodel.dtos.business.services.ProductTypeFactory
 
 /**
  * A DTO describing Product Identifiers
@@ -111,6 +113,40 @@ class ProductId implements Comparable<ProductId>{
      */
     String getType() {
         return type
+    }
+
+    /**
+     * Returns a ProductId from a given String representation
+     *
+     * Expects a ProductIdString with the Format P_N
+     * P being the one of the abbreviation values stored in {@link ProductType} enum
+     * N being an Integer Number
+     *
+     * @param String representation of a productId
+     * @return ProductId containing type and uniqueNumber of String representation
+     */
+    static ProductId from(String productId) {
+        if (!productId.contains("_")) {
+            throw new IllegalArgumentException("Not a valid product identifier.")
+        }
+        def splitId = productId.split("_")
+        String productTypeString
+        String runningNumberString
+        Long runningNumber
+        try {
+            ProductTypeFactory productTypeFactory = new ProductTypeFactory()
+            productTypeString = splitId[0].trim()
+            runningNumberString = splitId[1].trim()
+            runningNumber = Long.parseUnsignedLong(runningNumberString)
+            ProductType productType = productTypeFactory.getForString(productTypeString)
+        }
+        catch (NumberFormatException numberFormatException) {
+            throw new NumberFormatException("Provided productId does not have a valid uniqueID")
+        }
+        catch (IllegalArgumentException illegalArgumentException) {
+            throw new IllegalArgumentException("ProductId does not have a valid ProductType")
+        }
+        return new Builder(productTypeString, runningNumber).build()
     }
 
     /**